Move math operations to files
