+++ core/include/scipp/core/element/unary_operations.h
@@ -6,8 +6,6 @@
 
 #include <cmath>
 
-#include <Eigen/Dense>
-
 #include "scipp/common/overloaded.h"
 #include "scipp/core/element/arg_list.h"
 #include "scipp/core/transform_common.h"
@@ -17,37 +15,6 @@ namespace scipp::core {
 /// Operators to be used with transform and transform_in_place to implement
 /// operations for Variable.
 namespace element {
-
-constexpr auto abs = [](const auto x) noexcept {
-  using std::abs;
-  return abs(x);
-};
-
-constexpr auto abs_out_arg =
-    overloaded{arg_list<double, float>, [](auto &x, const auto y) {
-                 using std::abs;
-                 x = abs(y);
-               }};
-
-constexpr auto norm = overloaded{arg_list<Eigen::Vector3d>,
-                                 [](const auto &x) { return x.norm(); },
-                                 [](const units::Unit &x) { return x; }};
-
-constexpr auto sqrt = [](const auto x) noexcept {
-  using std::sqrt;
-  return sqrt(x);
-};
-
-constexpr auto dot = overloaded{
-    arg_list<Eigen::Vector3d>,
-    [](const auto &a, const auto &b) { return a.dot(b); },
-    [](const units::Unit &a, const units::Unit &b) { return a * b; }};
-
-constexpr auto sqrt_out_arg =
-    overloaded{arg_list<double, float>, [](auto &x, const auto y) {
-                 using std::sqrt;
-                 x = sqrt(y);
-               }};
 
 auto unit_check_and_assign = [](units::Unit &a, const units::Unit &b,
                                 const units::Unit &repl) {
@@ -69,7 +36,6 @@ constexpr auto nan_to_num =
                unit_check_and_return};
 
 constexpr auto nan_to_num_out_arg = overloaded{
-
     transform_flags::expect_all_or_none_have_variance,
     [](auto &x, const auto y, const auto &repl) {
       using std::isnan;
@@ -116,24 +82,6 @@ constexpr auto negative_inf_to_num_out_arg =
                    x = std::isinf(y) && y < 0 ? repl : y;
                },
                unit_check_and_assign};
-
-constexpr auto reciprocal = overloaded{
-    arg_list<double, float>,
-    [](const auto &x) {
-      return static_cast<
-                 core::detail::element_type_t<std::decay_t<decltype(x)>>>(1) /
-             x;
-    },
-    [](const units::Unit &unit) { return units::one / unit; }};
-
-constexpr auto reciprocal_out_arg = overloaded{
-    arg_list<double, float>,
-    [](auto &x, const auto &y) {
-      x = static_cast<core::detail::element_type_t<std::decay_t<decltype(y)>>>(
-              1) /
-          y;
-    },
-    [](units::Unit &x, const units::Unit &y) { x = units::one / y; }};
 
 } // namespace element
