+++ variable/CMakeLists.txt
@@ -9,6 +9,7 @@ set(INC_FILES
     include/scipp/variable/except.h
     include/scipp/variable/indexed_slice_view.h
     include/scipp/variable/logical.h
+    include/scipp/variable/math.h
     include/scipp/variable/misc_operations.h
     include/scipp/variable/operations.h
     include/scipp/variable/reduction.h
@@ -32,6 +33,7 @@ set(SRC_FILES
     except.cpp
     inplace_arithmetic.cpp
     logical.cpp
+    math.cpp
     operations.cpp
     rebin.cpp
     reduction.cpp
