+++ variable/include/scipp/variable/operations.h
@@ -6,6 +6,7 @@
 
 #include "scipp/variable/arithmetic.h"
 #include "scipp/variable/logical.h"
+#include "scipp/variable/math.h"
 #include "scipp/variable/misc_operations.h"
 #include "scipp/variable/reduction.h"
 #include "scipp/variable/shape.h"
